# openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "MediView API", "version": "1.0.0", "description": "手机端医学影像演示平台 API" },
  "servers": [{ "url": "http://localhost:8787/api" }],
  "paths": {
    "/health": { "get": { "summary": "健康检查", "responses": { "200": { "description": "服务正常" } } } },
    "/auth/login": { "post": { "summary": "登录", "responses": { "200": { "description": "登录成功" }, "401": { "description": "账号或密码错误" } } } },
    "/cases": { "get": { "summary": "案例列表", "parameters": [{ "name": "search", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "案例列表" } } } },
    "/cases/{id}": { "get": { "summary": "案例详情", "parameters": [{ "$ref": "#/components/parameters/CaseId" }], "responses": { "200": { "description": "案例详情" }, "404": { "description": "案例不存在" } } } },
    "/cases/{id}/plans": { "post": { "summary": "保存查看方案", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/CaseId" }], "responses": { "201": { "description": "方案已保存" } } } },
    "/cases/{id}/comments": { "post": { "summary": "发表评论", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/CaseId" }], "responses": { "201": { "description": "评论已创建" } } } },
    "/cases/{id}/share": { "post": { "summary": "生成只读分享", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/CaseId" }], "responses": { "201": { "description": "分享已创建" } } } },
    "/share/{token}": { "get": { "summary": "访问只读分享", "parameters": [{ "name": "token", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "分享案例" }, "410": { "description": "分享已失效" } } } }
  },
  "components": {
    "parameters": { "CaseId": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } },
    "schemas": { "Login": { "type": "object", "required": ["username", "password"], "properties": { "username": { "type": "string" }, "password": { "type": "string" } } } },
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer" } }
  }
}